also carry over body, and fix query parameters
